Work-in-progress

+++ Classes/parseCSV.m
@@ -90,7 +90,7 @@ static char *cstrstr(const char *haystack, const char needle) {
 	const char *allData = [_string cStringUsingEncoding:encoding];
 	if (allData == NULL )
 		return csvContent;
-	int length = strlen(allData);
+	int length = (int)strlen(allData);
 	[self setBufferSize:length+1];
 	char *textp, *laststop, *lineBeginning;
